refactor: colocate toggle group context ownership
