+++ packages/solid/src/toggle-group/context.ts
@@ -1,0 +1,5 @@
+import type { ToggleGroupStore } from "@tuiparts/core/toggle-group";
+import { createContext } from "solid-js";
+
+/** Private ToggleGroup ownership consumed by Toggle. */
+export const ToggleGroupContext = createContext<ToggleGroupStore>();
